refactor resourcemanager
